Update the save file

+++ src/graphics/renderer.cpp
@@ -78,7 +78,7 @@ auto renderer::update(const tile& tile, bool show_chunks, const camera& camera) 
         sand::from_hex(0xfad390)
     };
 
-    const auto camera_width = camera.zoom * (static_cast<float>(camera.screen_width) / camera.screen_height);
+    const auto camera_width = camera.zoom * (camera.screen_width / camera.screen_height);
     const auto camera_height = camera.zoom;
 
     if (d_texture.width() != camera_width || d_texture.height() != camera_height) {
